DONE wallet | minor Refactors

+++ blockchain/transaction.go
@@ -3,9 +3,9 @@ package blockchain
 import "encoding/json"
 
 type Transaction struct {
-	From   string `json:"from"`
-	To     string `json:"to"`
-	Amount uint64 `json:"amount"`
+	From   string  `json:"from"`
+	To     string  `json:"to"`
+	Amount float64 `json:"amount"`
 }
 
 type TransactionMessage struct {
